feat: add proper wrapping for esm and Deno by updating dependencies
- add esm test to npm run script
- run full tests from esm
