--- lib/string-utils.ts
@@ -1,30 +1,0 @@
-// Minimal replacement for ansi string helpers "wrap-ansi" and "strip-ansi".
-// to facilitate ESM and Deno modules.
-// TODO: look at porting https://www.npmjs.com/package/wrap-ansi to ESM.
-
-// The npm application
-// Copyright (c) npm, Inc. and Contributors
-// Licensed on the terms of The Artistic License 2.0
-// See: https://github.com/npm/cli/blob/4c65cd952bc8627811735bea76b9b110cc4fc80e/lib/utils/ansi-trim.js
-const ansi = new RegExp('\x1b(?:\\[(?:\\d+[ABCDEFGJKSTm]|\\d+;\\d+[Hfm]|' +
-'\\d+;\\d+;\\d+m|6n|s|u|\\?25[lh])|\\w)', 'g')
-
-export function stripAnsi (str: string) {
-  return str.replace(ansi, '')
-}
-
-export function wrap (str: string, width: number) {
-  const [start, end] = str.match(ansi) || ['', '']
-  str = stripAnsi(str)
-  let wrapped = ''
-  for (let i = 0; i < str.length; i++) {
-    if (i !== 0 && (i % width) === 0) {
-      wrapped += '\n'
-    }
-    wrapped += str.charAt(i)
-  }
-  if (start && end) {
-    wrapped = `${start}${wrapped}${end}`
-  }
-  return wrapped
-}
