+++ package.json
@@ -18,8 +18,7 @@
     "check": "standardx '**/*.ts' && standardx '**/*.js' && standardx '**/*.cjs'",
     "fix": "standardx --fix '**/*.ts' && standardx --fix '**/*.js' && standardx --fix '**/*.cjs'",
     "pretest": "rimraf build && tsc -p tsconfig.test.json && cross-env NODE_ENV=test npm run build:cjs",
-    "test": "c8 mocha ./test/*.cjs",
-    "test:esm": "c8 mocha ./test/esm/cliui-test.mjs",
+    "test": "c8 mocha ./test/cjs/*.cjs ./test/esm/*.mjs",
     "postest": "check",
     "coverage": "c8 report --check-coverage",
     "precompile": "rimraf build",
@@ -49,12 +48,15 @@
   "author": "Ben Coe <ben@npmjs.com>",
   "license": "ISC",
   "dependencies": {
-    "string-width": "^4.2.0",
-    "strip-ansi": "^6.0.1",
-    "wrap-ansi": "^7.0.0"
+    "string-width": "^5.1.2",
+    "strip-ansi": "^7.0.1",
+    "wrap-ansi": "^8.1.0"
   },
   "devDependencies": {
+    "@rollup/plugin-commonjs": "^25.0.0",
+    "@rollup/plugin-node-resolve": "^15.0.2",
     "@types/node": "^14.0.27",
+    "@types/wrap-ansi": "^8.0.1",
     "@typescript-eslint/eslint-plugin": "^4.0.0",
     "@typescript-eslint/parser": "^4.0.0",
     "c8": "^7.3.0",
@@ -67,8 +69,8 @@
     "gts": "^3.0.0",
     "mocha": "^10.0.0",
     "rimraf": "^3.0.2",
-    "rollup": "^2.23.1",
-    "rollup-plugin-ts": "^3.0.2",
+    "rollup": "^2.79.1",
+    "rollup-plugin-ts": "^3.2.0",
     "standardx": "^7.0.0",
     "typescript": "^4.0.0"
   },
