+++ rollup.config.js
@@ -1,4 +1,10 @@
 import ts from 'rollup-plugin-ts'
+import nodeResolve from '@rollup/plugin-node-resolve'
+import commonjs from '@rollup/plugin-commonjs'
+
+// These two transient dependencies are still CommonJS:
+// 'node_modules/emoji-regex/index.js',
+// 'node_modules/eastasianwidth/eastasianwidth.js'
 
 const output = {
   format: 'cjs',
@@ -12,6 +18,8 @@ export default {
   input: './lib/cjs.ts',
   output,
   plugins: [
-    ts({ /* options */ })
+    ts({ /* options */ }),
+    nodeResolve(),
+    commonjs()
   ]
 }
