+++ test/cjs/cliui-test.cjs
@@ -1,0 +1,4 @@
+const { runTests } = require('../shared-tests.cjs')
+const cliui = require('../../build/index.cjs')
+
+runTests(cliui, 'CJS')
